Update dari upstream

+++ app/src/main/java/uk/co/ribot/androidboilerplate/data/local/DatabaseHelper.java
@@ -1,6 +1,5 @@
 package uk.co.ribot.androidboilerplate.data.local;
 
-import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 
@@ -10,56 +9,41 @@ import com.squareup.sqlbrite.SqlBrite;
 import java.util.Collection;
 import java.util.List;
 
+import javax.inject.Inject;
+import javax.inject.Singleton;
+
 import rx.Observable;
 import rx.Subscriber;
 import rx.functions.Func1;
+import rx.schedulers.Schedulers;
 import uk.co.ribot.androidboilerplate.data.model.Ribot;
 
+@Singleton
 public class DatabaseHelper {
 
-    private BriteDatabase mDb;
+    private final BriteDatabase mDb;
 
-    public DatabaseHelper(Context context) {
-        mDb = SqlBrite.create().wrapDatabaseHelper(new DbOpenHelper(context));
+    @Inject
+    public DatabaseHelper(DbOpenHelper dbOpenHelper) {
+        SqlBrite.Builder briteBuilder = new SqlBrite.Builder();
+        mDb = briteBuilder.build().wrapDatabaseHelper(dbOpenHelper, Schedulers.immediate());
     }
 
     public BriteDatabase getBriteDb() {
         return mDb;
     }
 
-    /**
-     * Remove all the data from all the tables in the database.
-     */
-    public Observable<Void> clearTables() {
-        return Observable.create(new Observable.OnSubscribe<Void>() {
-            @Override
-            public void call(Subscriber<? super Void> subscriber) {
-                BriteDatabase.Transaction transaction = mDb.newTransaction();
-                try {
-                    Cursor cursor = mDb.query("SELECT name FROM sqlite_master WHERE type='table'");
-                    while (cursor.moveToNext()) {
-                        mDb.delete(cursor.getString(cursor.getColumnIndex("name")), null);
-                    }
-                    cursor.close();
-                    transaction.markSuccessful();
-                    subscriber.onCompleted();
-                } finally {
-                    transaction.end();
-                }
-            }
-        });
-    }
-
     public Observable<Ribot> setRibots(final Collection<Ribot> newRibots) {
         return Observable.create(new Observable.OnSubscribe<Ribot>() {
             @Override
             public void call(Subscriber<? super Ribot> subscriber) {
+                if (subscriber.isUnsubscribed()) return;
                 BriteDatabase.Transaction transaction = mDb.newTransaction();
                 try {
                     mDb.delete(Db.RibotProfileTable.TABLE_NAME, null);
                     for (Ribot ribot : newRibots) {
                         long result = mDb.insert(Db.RibotProfileTable.TABLE_NAME,
-                                Db.RibotProfileTable.toContentValues(ribot.profile),
+                                Db.RibotProfileTable.toContentValues(ribot.profile()),
                                 SQLiteDatabase.CONFLICT_REPLACE);
                         if (result >= 0) subscriber.onNext(ribot);
                     }
@@ -78,7 +62,7 @@ public class DatabaseHelper {
                 .mapToList(new Func1<Cursor, Ribot>() {
                     @Override
                     public Ribot call(Cursor cursor) {
-                        return new Ribot(Db.RibotProfileTable.parseCursor(cursor));
+                        return Ribot.create(Db.RibotProfileTable.parseCursor(cursor));
                     }
                 });
     }
